verify routes with postman, test on joi, ok, test on token, ok.
