+++ server/src/types/express/index.d.ts
@@ -1,4 +1,3 @@
-// to make the file a module and avoid the TypeScript error
 export type {};
 
 declare global {
@@ -24,12 +23,14 @@ declare global {
     password: string;
   }
 
+  export type MyPayload = JwtPayload & { sub: string };
+
   namespace Express {
     export interface Request {
       /* ************************************************************************* */
       // Add your custom properties here, for example:
       //
-      // user?: { ... }
+      auth: MyPayload;
       /* ************************************************************************* */
     }
   }
